pyproject.toml
[project]
name = "ai-knowledge-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "camelot-py[cv]>=1.0.9",
    "google-generativeai>=0.8.0",
    "ipykernel>=7.0.1",
    "openai>=1.0.0",
    "pre-commit>=4.3.0",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "qdrant-client>=1.12.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-qdrant>=0.2.0",
    "langchain-community>=0.3.0",
    "sentence-transformers>=3.3.0",
    "rank-bm25>=0.2.2",
    "einops>=0.8.0",
    "pydantic>=2.10.0",
    "requests>=2.31.0",
    "nltk>=3.9.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "scikit-learn>=1.7.2",
    "joblib>=1.5.2",
    "transformers>=4.57.1",
    "datasets>=4.3.0",
    "torch>=2.9.0",
    "pandas>=2.3.3",
    "accelerate>=1.11.0",
    "safetensors>=0.4.0",
]
